commit by rb

+++ IBS/Models/CallRegisterModel.cs
@@ -63,11 +63,11 @@ namespace IBS.Models
         [Required(ErrorMessage = "Please enter a Qty To Insp")]
         public decimal? QtyToInsp { get; set; }
 
-        public int? QtyPassed { get; set; }
+        public decimal? QtyPassed { get; set; }
 
-        public int? QtyRejected { get; set; }
+        public decimal? QtyRejected { get; set; }
 
-        public int? QtyDue { get; set; }
+        public decimal? QtyDue { get; set; }
 
         public string? Consignee { get; set; }
